Add a readme viewer component
