+++ assets/stylesheets/_components.scss
@@ -194,6 +194,7 @@
 @import 'components/reader-main/style';
 @import 'components/reader-infinite-stream/style';
 @import 'components/reader-popover/style';
+@import 'components/readme-viewer/style';
 @import 'components/rewind-credentials-form/style';
 @import 'components/ribbon/style';
 @import 'components/screen-reader-text/style';
